Connect form component to store and move to container folder
